test(use-ws-runtime): consolidate parallel test suites (#313)
Merge the 44 tests from `src/hooks/__tests__/use-ws-runtime.test.ts` into
the canonical `src/__tests__/hooks/use-ws-runtime.test.ts`, adopting the
real-WebSocket mock + identity `useExternalStoreRuntime` strategy
throughout (tests now call `result.current.runtime.onNew(...)` and read
`runtime.messages` instead of capturing callbacks via module-scope vars).

Removes the foot-gun where every new module mock had to be mirrored
across two files. Adds `simulateOpen()`, `simulateMessage()`, and a
top-level `latestWs()` helper to the canonical file so migrated blocks
stay readable.

All 126 tests preserved, full web suite green (4370 passed).
